last changes for db

+++ Oprim.Domain/Common/BaseEntity.cs
@@ -13,7 +13,7 @@ public class BaseEntity
     // public bool IsDelete { get; set; }= false;
     
     [ForeignKey("CreatorId")] public Stakeholder? Creator { get; set; }
-    public int CreatorId { get; set; }
+    public long CreatorId { get; set; }
 
     [MaxLength(20)] public DateTime CreateTime { get; set; }
 }
